// ts/Series/Column/ColumnMetricsObject.d.ts
/* *
 *
 *  (c) 2010-2021 Torstein Honsi
 *
 *  License: www.highcharts.com/license
 *
 *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
 *
 * */

/* *
 *
 *  Declarations
 *
 * */

interface ColumnMetricsObject {
    offset: number;
    width: number;
    paddedWidth: number;
    columnCount: number;
}

/* *
 *
 *  Default Export
 *
 * */

export default ColumnMetricsObject;