// ts/Core/Renderer/SizeObject.d.ts
/* *
 *
 *  (c) 2010-2021 Torstein Honsi
 *
 *  License: www.highcharts.com/license
 *
 *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
 *
 * */

/* *
 *
 *  Declarations
 *
 * */

export interface SizeObject {
    height: number;
    width: number;
}

/* *
 *
 *  Default Export
 *
 * */

export default SizeObject;